+++ EZOCore.lua
@@ -2,8 +2,8 @@ EZOCore = EZOCore or {}
 local EZOCore = EZOCore
 
 EZOCore.name = "EZOCore"
-EZOCore.version = "0.1.15"
-EZOCore.addOnVersion = 115
+EZOCore.version = "0.1.16"
+EZOCore.addOnVersion = 116
 EZOCore.apiVersion = 1
 
 EZOCore.EVENT_INITIALIZED = "EZO_CORE_INITIALIZED"